refactor: split catalog inventory builder in collector modulari
- Crea scripts/collectors/ con base.py, ckan.py, sdmx.py, sparql.py
- Riduce build_catalog_inventory.py a orchestratore ~120 righe
- Aggiorna test: monkeypatch spostato su collectors.ckan/sparql
- Registry: inps sample_size 25→100, openbdap e lavoro_opendata +package_show_sample
- CI: --workers=4

Closes #106
